Add the ietf-ip YANG model (RFC 7277)
While the ietf-ip YANG model has a newer revision in RFC 8344, this
revision is used in more than one ODL project (sfc, honeycomb/vbd), each
one providing their own file, which is not a good practice. The model
should live in MD-SAL, and then the other project depend on it.

Change-Id: I8cc5ea99f16ae9bcc52afca9436fe5251a4d6361
Signed-off-by: Lorand Jakab <lojakab@cisco.com>