Rename Identifiable.getKey() to key()
Identifiable's getKey() method has a tendency to conflict with models
which define a leaf named 'key'. Rename the method so it does not
conflict with generated getters and unify its definition in
BindingMapping. Also tag it as returning @NonNull, as that is the API
contract.

Since 'key' is a valid property, Builders cannot use a 'setKey' method,
hence it is renamed to 'withKey'.

JIRA: MDSAL-2
Change-Id: Idda720c31fd5519ab21a1722ab5072da5b18fcc9
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>