work around suspected concurrency bug in AbstractStreamWriterGenerator
by retrying (only once) a Cache get, because if we are hit by the
"frozen class" situation, then (somehow; unclear how this is possible) a
parallel thread may have created the requested
DataObjectSerializerImplementation in the mean time, and it should
succeed now.

This is a hack, and I would prefer a clean solution, but I don't yet see
the concurrency issue from looking at the CacheLoader use in
AbstractStreamWriterGenerator - that seems fine.

JIRA: MDSAL-213
Change-Id: Ibfb459e0a9617a59d2d9c67aa4f80b51267902a9
Signed-off-by: Michael Vorburger <vorburger@redhat.com>