Teach InstanceIdentifier about case-based addressing
InstanceIdentifier needs to be able to address items which are
coming from a grouping referenced by uses immediately within
a case statement.

The reason for this is that the items in the grouping do not have
their namespace capture class generated for the instantiation,
hence they are not addressable via the usual ChildOf<> relationship.

With the advent of ChoiceIn<>, we can use the classes generated
for case statements to act as a namespace anchor and pivot to
the grouping member. This patch adds the necessary methods.

JIRA: MDSAL-45
Change-Id: I1cc626e391b450e5b32c505ecf6021c39f55532d
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>