make validate/prepare/commit methods in InMemoryDOMDataStore protected
instead of package private; because I'd like to @Override them in
something I am experimenting with (a POC where I would like to extend
InMemoryDOMDataStore to send the DataTreeCandidate to an external data
store on commit).

This is the equivalent of controller's I5a4e0ca0c1564df838a1faf38e951ae7010affc6.

Change-Id: I877ef3bc7ea02e9366ea93cf82d8660c1fba135d
Signed-off-by: Michael Vorburger <vorburger@redhat.com>