Add Binding{Data,RpcFuture}Aware bridge interfaces
We need to be able to communicate the binding future between two
incarnations of LazyDOMRpcResultFuture: controller and mdsal.

Create a migration interface for this purpose, which exposes
the required method. Convert RpcServiceAdapter to use it instead
of relying on LazyDOMRpcResultFuture class.

The same goes for LazySerializedContainerNode, bridged via
BindingDataAware.

JIRA: SFC-222
Change-Id: Icf5df0a2cfb4ec24d404ac6d077c2755692b9d5c
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>