Update patch set 1
Patch Set 1:

> Is this really needed? xtend is used solely in compile time and it
 > works as-is.

It happens to work as-is because our Nexus has the right combination of artifacts; but the constraints currently specified in our POM aren’t strict enough to fix the Xtend breakage in all cases, and it is possible to end up with an artifact setup which satisfies the constraints but fails when the plugin runs.

This patch specifies the constraints strictly enough to fail at resolution time (and more importantly, forces Maven to download artifacts which do meet the real requirements).

Patch-set: 1
Reviewer: Gerrit User 4540 <4540@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0