Optimize Ip{Address,Prefix}Builder
These builders are heavily used and are overly defensive, to the point
where they do not follow YANG specification -- which clearly states that
members need to be tried in the order of declaration.

This means there is no ambiguity between ipv4 and ipv6 and if IPv4
succeeds, that's what we'll use. This allows us to reduce number
of regex matches performed, which retaining correctness.

Also remove ugly duplication of patterns are pick them up from the pattern
constants in the class.

Change-Id: Ia35d040881e721375b240dd5848ad9501ae2e192
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>