move AbstractDataBrokerTest & Co. to test-utils; ditch adapter test-jar
because the mdsal-binding-dom-adapter test-jar was a PITA (for
transitive dependency management); now downstream must replace their
dependencies to it by mdsal-binding-test-utils.

All the mdsal-binding-dom-adapter test classes are moved to a new 
mdsal-binding-dom-adapter-tests project because mdsal-binding-test-utils
needs to depend on (main) mdsal-binding-dom-adapter, so tests (in
mdsal-binding-dom-adapter) cannot depend on mdsal-binding-test-utils,
otherwise there would be a circular dependency.

Change-Id: I1a3ddfbe5c09fa336e638fd11ac70d1eb5b8e012
Signed-off-by: Michael Vorburger <vorburger@redhat.com>