Implement managed transactions
Managed transactions ensure that transactions are always closed in the
appropriate way:
* read-only transactions are closed when they are no longer needed;
* read-write and write-only transactions are submitted if the code
  manipulating them didn’t throw an exception, cancelled otherwise.

Managed transactions are datastore-specific and enforce either
configuration or operational reads and writes.

Two providers are included: a basic transaction manager which provides
transactions, ensures they are closed, without any extras, and a
retrying transaction manager which will retry if a retriable error
occurs (i.e. an OptimisticLockFailedException when writing, or a
ReadFailedException when reading).

This is copied from Genius where it’s used extensively; it’s also used
in NetVirt.

Change-Id: I391ca4ca50d1b1177d827e33c0d11e98378e1fb4
Signed-off-by: Stephen Kitt <skitt@redhat.com>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>