Split off Read/WriteOperations from Read/WriteTransaction
It is oftentimes useful to allow access to a transaction's operations
without allowing the state of the transaction to be modified. This
patch splits off operations from transactions, so proper isolation
can be maintained.

This also resolves the issue of ReadWriteTransaction having both
cancel() and close() methods -- since it has a write aspect, it
inherits lifecycle methods from WriteTransaction.

Change-Id: I98c02c0c5cbcac2432f629973860c56cf2818d6b
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>