Update patch set 1
Patch Set 1:

Ugh, that probably is an omission. When a subshard is attached, existing producers need to be updated to account for that subshard (if their path includes the subshard). This can be quite tricky, as there may be transactions being built up, etc.

The simplest we can do here is to fail affected producers and let the user re-create them with the new layout. Not perfect, but sharding layout changes should be very rare and this way we avoid a ton of complexity.

Patch-set: 1