WIP: Populate augmentation cache on single access
This implements an intermediate single-item cache when an augmentation
is fetched. In order to be able to discern whether a cache is a partial
one or a fully-resolved one we use Collections.singletonMap() for
partial case and ImmutableMap for a full cache -- hence the check is
a simple instanceof.

Single cache is expanded to a full cache when a second augmentation
is accessed.

As cachedAugmentations can now be in three states null, single, full,
we need to deal with more race conditions happening, notably
following transitions are possible:
- null -> single (race: another full or another single)
- null -> full (race: another full or another single)
- single -> full (race: another full)

Change-Id: Ic70ec051cfe3611c30eae0cc205bb220b088189d
JIRA: MDSAL-377
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>