Return empty list for non-present non-mandatory lists
The data store removes lists which become empty, which leads
to a lot of lists being null. Rather than forcing all code to
deal with that, change semantics of lists which are not constrained
by min-elements > 0 so that non-present lists are reported as
empty lists.

Lists which have a min-elements > 0 get still reported as null,
so we do not return a List which has fewer elements than the minimum
required by the model.

Change-Id: Iedfa86f455c5889702d91f9a2777983d5302bc87
JIRA: MDSAL-18
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>