Update patch set 1
Patch Set 1:

I am not sure about the placement, but this is the lowest place which is not yang-binding. I do not want to overload yang-binding more, not until we move those things to a proper namespace.

Patch-set: 1
Reviewer: Gerrit User 649 <649@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0