Make nullToEmpty return a modifiable list
In ODL, YANG lists are modifiable; nullToEmpty() is designed to wrap
potentially absent YANG lists, so it might make sense to make it
return a modifiable list. For the basic foreach use-case documented in
the Javadocs, an immutable list is fine, but code which manipulates
the resulting list currently assumes it can modify it.

Change-Id: I0c66b2645a6d6fba165e3ac56a0f6c54225b008e
Signed-off-by: Stephen Kitt <skitt@redhat.com>