Update patch set 13
Patch Set 13: Code-Review-1

I think we need to revise the entire idea of producer exclusivity. It makes interaction between RESTCONF and applications problematic (i.e. if an app writes to the config datastore, restconf won't ever be able to interfere) and there actually are users who use the overlap -- making the migration a huge pain.

On clustering side of things, exclusivity does not quite work with partitions and the fact that producer allocation is synchronous: in cluster we need to ensure cluster-wide sync on what producers are there and need to eject disconnected nodes.

Can you open a JIRA for tracking that effort, please?

Patch-set: 13
Label: Code-Review=-1