Update patch set 23
Patch Set 23: Code-Review+1

> Tom, Stephen: any objections on current naming? I'd like to get
 > mdsal-3.0.2 out early next week...

I don’t particularly like it but I don’t have anything better to offer, so let’s get it in. Perhaps a @Beta annotation would be in order, so we’re not stuck with it? With a comment saying that it will be set in stone for 4.0.0 ;-).

Patch-set: 23
Reviewer: Gerrit User 4540 <4540@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1