WIP: Use MethodHandles to invoke accessors
This is a two-aspect patch, switching NodeContextSupplier to
string-based lookup and using MethodHandles to invoke accessors
on object being compared.

Since we no longer rely on byMethod.keySet() to provide us with
Methods in hashCode order, we dial down the implementation to
an ImmutableMap.

What we are saving on the instantiated keySet(), we are expeding
on allocating two arrays, one for method names and the other one
for MethodHandles, which retain hashCode order. These arrays
are then package-private-leaked to LazyDataObject, which uses
them to implement hashCode()/equals()/toString().

FIXME: evaluate memory footprint/performance changes with this
       patch.

JIRA: MDSAL-398
JIRA: MDSAL-399
Change-Id: Ie18c4df13f11d1ddcf1b099289a463fe6ebf06f3
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>