Update patch set 24
Patch Set 24: Code-Review+2

Right, but that does not really express the difference -- a put() of a list entry traversing multiple non-presence containers will succeed, no matter what. put() of a leaf crossing a non-existent list entry.

We really want that method to be used only when necessary, as it really is a lot less performant and detecting when it is provably not needed requires human thinking.

Patch-set: 24