use ImmutableMap.Builder in IdentifiableItemCodec serialize()
This is more efficient than the LinkedHashMap, even with
initialCapacity.  Unfortunately yangtools' ImmutableOffsetMap which is
constructed in the YangInstanceIdentifier.NodeIdentifierWithPredicates
from this Map will still make a copy, which perhaps in an ideal world in
the future could even also be avoided; but doing this optimization here
at least saves a wasteful LinkedHashMap.

JIRA: MDSAL-387
Change-Id: If4d1038238ac5065669924adf9de3a2eb8d6e76b
Signed-off-by: Michael Vorburger <vorburger@redhat.com>