Generate implementedInterface
getImplementedInterface() is trampling on getter namespace, making
it possible to clash with a user-supplied model. It also is not properly
overridden in generated code.

Fix both these issues by introducing DataContainer.implementedInterface()
and overriding it as needed wither with a narrower specification
(groupings) or a default implementation (other interfaces).

JIRA: MDSAL-396
Change-Id: I902350f0979067a9e035770e683ab18e966deb7f
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>