Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

I still think we need to change it to onEmptyInitialData or onInitialDataComplete - onInitialData is really wrong. I understand technically it's an API contract but it was only added a month and 1/2 ago and nobody is using yet nor has it been implemented so it would be nice to just change it and release a new mdsal version.  Otherwise I say we add the new method as default and deprecate onInitialData.

Patch-set: 3
Reviewer: Gerrit User 1842 <1842@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=-1