ditch infrautils.inject.guice from new wiring/ artifact
The AutoWiringModule and GuiceClassPathBinder are actually complete
overkill in this particular case; it's not needed, as we already
explicitly bind what we need anyway.. this will be clearer, have better
error message, and faster to initialize.

This also means that the wiring/ artifact depends on infrautils only for
the GuiceRule for the test, and at run-time it only depends on pure
Guice - and even that only optionally.

JIRA: MDSAL-418
Change-Id: I563e4aa1056bb200dee653112f9ee47f4656ff2d
Signed-off-by: Michael Vorburger <vorburger@redhat.com>