Add EventInstantAware interface
This interface will be used to carry the equivalent of RFC5277
notification eventTime. It is modeled exactly the same way as
DOMEvent/DOMNotification, hence it will allow us to bridge the gap
in binding-dom-adapter, where we are losing this information.

JIRA: MDSAL-282
Change-Id: Ie93015c2609a2d1c59df6c63f76511aa68867832
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit 992122145777924ad5b3a52da65188d8a91b8246)