WIP: create a separate classloader for holding codec objects
In order to make our code simpler and more performant, we can
turn the use of java.lang.reflect.Proxy into generated classes,
but in order to make things work correctly, we need to make sure
we do not pollute the classpath.

This patch introduces CodecClassLoader, which is a separate
classloader, maintaining a connection between the codec class loader
and classloaders of individual bindings. It also brings in
Javassist, which we use to generate classes.

As an example, ForeignOpaqueObject is converted from being a proxy
to being an abstract base class, over which a concrete implementation
is generated.

This approach has a few advantages:
- we do not generate a proxy class in public space, hence our mess
  will undergo normal GC rules
- generated code is a normal final class, hence we can inherit
  method implementations from interfaces and subclasses
- for each instance, we do not require a proxy and an invocation
  handler, so we instantiate only one object
- there is no reflection in the invocation path, hence JIT has
  full (and direct) visibility and can perform all of its magic

// FIXME: this needs to be tested in a multi-classloader environment

Change-Id: I2be5bb88f9557375f517415d6bec9f68e803628c
JIRA: MDSAL-442
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>